requirements.txt
numpy==1.26.4
tqdm==4.66.4
trimesh==4.4.0
pyvista==0.45.2   
tensorboard==2.16.2
warp-lang==1.1.0
anthropic==0.26.1
mistralai==0.3.0
openai==1.62.0
opencv-python==4.9.0.80 
e3nn==0.5.1 
einops==0.8.0 
natsort==8.4.0 
nerfview==0.0.3 
omegaconf==2.3.0 
pyvista==0.44.0 
splines==0.3.2 
torchmetrics 
tensorboardX 
tqdm 
plyfile 
trimesh 
mediapy 
imageio 
viser==0.2.3
